release: decrease log level for preflight check
Previously, the preflight check was logging at the trace level, which
created a huge log file, what caused the API server to reject
submissions due to the log file size. At the same time, the catalog
still showed the image as published, which was misleading. This commit
reduce the log level to error.

Release note: none
Epic: none